+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 All notable changes to this project will be documented in this file. Format based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Latest]
+- Fix **ContainerFreeForm** the largest empty area computation
 
 ## [7.8.0] 2023 08 06
 - Update Java to 20